chore: rename lib/gui/env_vars.h => lib/gui/EnvVars.h
